+++ src/dawn/native/ComputePassEncoder.cpp
@@ -469,7 +469,8 @@ namespace dawn::native {
                 if (offsets.empty()) {
                     APISetBindGroup(static_cast<uint32_t>(i), bg);
                 } else {
-                    APISetBindGroup(static_cast<uint32_t>(i), bg, offsets.size(), offsets.data());
+                    APISetBindGroup(static_cast<uint32_t>(i), bg,
+                                    static_cast<uint32_t>(offsets.size()), offsets.data());
                 }
             }
         }
